use %ld and %lu string specifiers for NSInteger and NSUInteger
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@15140 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BDSKBookmarkController.m
@@ -437,7 +437,7 @@ static id sharedBookmarkController = nil;
     } else if ([tcID isEqualToString:@"url"]) {
         if ([item bookmarkType] == BDSKBookmarkTypeFolder) {
             NSInteger count = [[item children] count];
-            return count == 1 ? NSLocalizedString(@"1 item", @"Bookmark folder description") : [NSString stringWithFormat:NSLocalizedString(@"%i items", @"Bookmark folder description"), count];
+            return count == 1 ? NSLocalizedString(@"1 item", @"Bookmark folder description") : [NSString stringWithFormat:NSLocalizedString(@"%ld items", @"Bookmark folder description"), (long)count];
         } else {
             return [item urlString];
         }
