+++ BDSKRISParser.m
@@ -193,10 +193,10 @@
         } else if([key isSingleValuedField] || [key isURLField]) {
             // for single valued and URL fields, create a new field name
             NSInteger i = 1;
-            NSString *newKey = [key stringByAppendingFormat:@"%d", i];
+            NSString *newKey = [key stringByAppendingFormat:@"%ld", (long)i];
             while ([pubDict objectForKey:newKey] != nil) {
                 i++;
-                newKey = [key stringByAppendingFormat:@"%d", i];
+                newKey = [key stringByAppendingFormat:@"%ld", (long)i];
             }
             key = newKey;
             newString = [value copy];
