+++ BDSKSearchBookmarkController.m
@@ -336,7 +336,7 @@ static BDSKSearchBookmarkController *sharedBookmarkController = nil;
     } else if ([tcID isEqualToString:@"server"]) {
         if ([item bookmarkType] == BDSKSearchBookmarkTypeFolder) {
             NSInteger count = [[item children] count];
-            return count == 1 ? NSLocalizedString(@"1 item", @"Bookmark folder description") : [NSString stringWithFormat:NSLocalizedString(@"%i items", @"Bookmark folder description"), count];
+            return count == 1 ? NSLocalizedString(@"1 item", @"Bookmark folder description") : [NSString stringWithFormat:NSLocalizedString(@"%ld items", @"Bookmark folder description"), (long)count];
         } else {
             return [[item info] valueForKey:@"name"];
         }
